refactor: loggin encapsulated
