+++ pyproject.toml
@@ -30,21 +30,47 @@ classifiers = [
 dependencies = [
     # Core framework
     "hydra-core>=1.3.2",
-    "mlflow>=3.9.0",
     "numpy>=2.4.2",
     # Deep learning (vision)
     "torch>=2.10.0",
     "torchvision>=0.20.0",
     "pillow>=10.0.0",
     "pandas>=2.0.0",
-    # Transparency/explainability
-    "shap>=0.46.0",
-    "captum>=0.7.0",
     # Visualization
     "matplotlib>=3.9.0",
-    # Performance Metrics
+]
+
+[project.optional-dependencies]
+mlflow = [
+    "mlflow>=3.9.0",
+]
+shap = [
+    "shap>=0.46.0",
+]
+captum = [
+    "captum>=0.7.0",
+]
+metrics = [
     "torchmetrics>=1.8.2",
     "faster-coco-eval>=1.6.3",
+]
+all = [
+    "mlflow>=3.9.0",
+    "shap>=0.46.0",
+    "captum>=0.7.0",
+    "torchmetrics>=1.8.2",
+    "faster-coco-eval>=1.6.3",
+]
+
+[dependency-groups]
+dev = [
+    "ruff>=0.8.0",
+    "pre-commit>=4.0.0",
+    "commitizen>=3.0.0",
+    "pyright>=1.1.390",
+    "pytest>=8.0.0",
+    "pytest-cov>=4.1.0",
+    "pycocotools>=2.0.10",
 ]
 
 [project.scripts]
@@ -57,17 +83,6 @@ Homepage = "https://github.zhaw.ch/RAI/Tech-Assessment-Platform"
 Repository = "https://github.zhaw.ch/RAI/Tech-Assessment-Platform.git"
 Issues = "https://github.zhaw.ch/RAI/Tech-Assessment-Platform/issues"
 Source = "https://github.zhaw.ch/RAI/Tech-Assessment-Platform"
-
-[dependency-groups]
-dev = [
-    "ruff>=0.8.0",
-    "pre-commit>=4.0.0",
-    "commitizen>=3.0.0",
-    "pyright>=1.1.390",
-    "pytest>=8.0.0",
-    "pytest-cov>=4.1.0",
-    "pycocotools>=2.0.10",
-]
 
 [build-system]
 requires = ["hatchling"]
