+++ src/raitap/configs/tracking/mlflow.yaml
@@ -1,4 +1,3 @@
-enabled: true
 tracking_uri: http://127.0.0.1:5000
 registry_uri: null
 log_model: false
