+++ src/raitap/metrics/classification_metrics.py
@@ -4,11 +4,13 @@ from typing import TYPE_CHECKING, Any, Literal
 
 from torchmetrics import Accuracy, F1Score, Precision, Recall
 
-from .base import MetricComputer, MetricResult
+from .base_metric import MetricComputer, MetricResult
 from .utils import tensor_to_python
 
 if TYPE_CHECKING:
     import torch
+
+    from raitap.tracking.base_tracker import BaseTracker
 
 Task = Literal["binary", "multiclass", "multilabel"]
 Average = Literal["micro", "macro", "weighted", "none"]
@@ -115,3 +117,6 @@ class ClassificationMetrics(MetricComputer):
         self.precision.reset()
         self.recall.reset()
         self.f1.reset()
+
+    def log(self, tracker: BaseTracker) -> None:
+        tracker.log_metrics(self.compute().metrics, prefix="performance")
