--- src/raitap/tracking/base.py
@@ -1,20 +1,0 @@
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Any, Protocol
-
-if TYPE_CHECKING:
-    from pathlib import Path
-
-
-class Tracker(Protocol):
-    def start_assessment(self, assessment_name: str) -> None: ...
-    def log_config(self, config: Any) -> None: ...
-    def log_model(self, model: Any, artifact_path: str = "model") -> None: ...
-    def log_dataset(self, dataset_info: dict[str, Any], artifact_path: str = "dataset") -> None: ...
-    def log_artifacts(self, local_dir: str | Path, artifact_path: str) -> None: ...
-    def log_metrics(
-        self,
-        metrics: dict[str, float | int | bool],
-        prefix: str = "performance",
-    ) -> None: ...
-    def finalize(self, status: str = "FINISHED") -> None: ...
